Fix zabbix-agent service on Lucid
Without this patch Puppet does not correctly manage the zabbix agent
because the return code is not accurate.  The zabbix-agent init script
looks to the wrong directory for the PID file, /var/run/zabbix-agent is
incorrect and the PID is actually in /var/run/zabbix/zabbix_agentd.pid

This patch fixes the problem by managing a correct init script.