Bounce zabbix-agent when doing an initial install
The zabix::agent class was getting the agent into a state where Puppet
thought it was not running and could not start the service because the
process is in fact running.

This problem came about because we're replacing the init script in the
first puppet run and the path to the PID file changes.

This patch fixes the problem by stopping the agent in an exec resource
if we replace the init script.